Switch to eine/setup-msys2 to use the MSYS2 environment
Current Action containers contain a MSYS2 environment which is installed at the location
the setup script expects. The original numworks/setup-msys2 Action can only use a temporary
environment, the solution to create a symlink at the default location name to it
is not possible anymore. The now used fork can use the installed environment,
this requires however to not use the feature to install a release version of it.